[GraphQL][MVR] Move to new naming standard (#19376)
Move things left and right (gets rid of the `dot_move` naming), and
update our parsing logic to use our new naming standard.

The naming standard used is `{ns_name}/{app_name}/{version}` with
`/{version}` being optional.

Tweaked the existing e2e & unit tests to work with the new naming
standard, and resolution works as expected on the e2e ones.

---

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK:
- [ ] REST API:
