+++ crates/sui-graphql-rpc/src/error.rs
@@ -5,7 +5,7 @@ use async_graphql::{ErrorExtensionValues, ErrorExtensions, Pos, ServerError};
 use sui_indexer::errors::IndexerError;
 use sui_json_rpc::name_service::NameServiceError;
 
-use crate::types::dot_move::error::MoveRegistryError;
+use crate::types::move_registry::error::MoveRegistryError;
 
 /// Error codes for the `extensions.code` field of a GraphQL error that originates from outside
 /// GraphQL.
