+++ crates/sui-graphql-rpc/src/types/mod.rs
@@ -16,7 +16,6 @@ pub(crate) mod datatype;
 pub(crate) mod date_time;
 pub(crate) mod digest;
 pub(crate) mod display;
-pub(crate) mod dot_move;
 pub(crate) mod dry_run_result;
 pub(crate) mod dynamic_field;
 pub(crate) mod epoch;
@@ -30,6 +29,7 @@ pub(crate) mod move_function;
 pub(crate) mod move_module;
 pub(crate) mod move_object;
 pub(crate) mod move_package;
+pub(crate) mod move_registry;
 pub(crate) mod move_struct;
 pub(crate) mod move_type;
 pub(crate) mod move_value;
