+++ crates/sui-graphql-rpc/src/types/move_registry/on_chain.rs
@@ -8,6 +8,7 @@ use move_core_types::language_storage::StructTag;
 use once_cell::sync::Lazy;
 use regex::Regex;
 use serde::{Deserialize, Serialize};
+use sui_json_rpc::name_service::{validate_label, Domain};
 use sui_types::{
     base_types::{ObjectID, SuiAddress},
     collection_types::VecMap,
@@ -23,33 +24,27 @@ use crate::{
 
 use super::error::MoveRegistryError;
 
-const MAX_LABEL_LENGTH: usize = 63;
-
 /// Regex to parse a dot move name. Version is optional (defaults to latest).
-/// For versioned format, the expected format is `app@org/v1`.
-/// For an unversioned format, the expected format is `app@org`.
+/// For versioned format, the expected format is `@org/app/1` (1 == version).
+/// For an unversioned format, the expected format is `@org/app`.
 ///
 /// The unbound regex can be used to search matches in a type tag.
 /// Use `VERSIONED_NAME_REGEX` for parsing a single name from a str.
-const VERSIONED_NAME_UNBOUND_REGEX: &str = concat!(
-    "([a-z0-9]+(?:-[a-z0-9]+)*)",
-    "@",
-    "([a-z0-9]+(?:-[a-z0-9]+)*)",
-    r"(?:\/v(\d+))?"
-);
+const VERSIONED_NAME_UNBOUND_REGEX: &str =
+    concat!(r"([a-z0-9.\-@]*)", r"\/", "([a-z0-9.-]*)", r"(?:\/(\d+))?",);
 
 /// Regex to parse a dot move name. Version is optional (defaults to latest).
-/// For versioned format, the expected format is `app@org/v1`.
-/// For an unversioned format, the expected format is `app@org`.
+/// For versioned format, the expected format is `@org/app/1`.
+/// For an unversioned format, the expected format is `@org/app`.
 ///
 /// This regex is used to parse a single name (does not do type_tag matching).
 /// Use `VERSIONED_NAME_UNBOUND_REGEX` for type tag matching.
 const VERSIONED_NAME_REGEX: &str = concat!(
     "^",
-    "([a-z0-9]+(?:-[a-z0-9]+)*)",
-    "@",
-    "([a-z0-9]+(?:-[a-z0-9]+)*)",
-    r"(?:\/v(\d+))?",
+    r"([a-z0-9.\-@]*)",
+    r"\/",
+    "([a-z0-9.-]*)",
+    r"(?:\/(\d+))?",
     "$"
 );
 
@@ -57,7 +52,7 @@ const VERSIONED_NAME_REGEX: &str = concat!(
 pub(crate) static VERSIONED_NAME_UNBOUND_REG: Lazy<Regex> =
     Lazy::new(|| Regex::new(VERSIONED_NAME_UNBOUND_REGEX).unwrap());
 
-/// A regular expression that detects a single name in the format `app@org/v1`.
+/// A regular expression that detects a single name in the format `@org/app/1`.
 pub(crate) static VERSIONED_NAME_REG: Lazy<Regex> =
     Lazy::new(|| Regex::new(VERSIONED_NAME_REGEX).unwrap());
 
@@ -67,6 +62,7 @@ pub(crate) static VERSIONED_NAME_REG: Lazy<Regex> =
 #[derive(Debug, Serialize, Deserialize, Clone, Eq, PartialEq)]
 pub(crate) struct AppRecord {
     pub(crate) app_cap_id: ID,
+    pub(crate) ns_nft_id: ID,
     pub(crate) app_info: Option<AppInfo>,
     pub(crate) networks: VecMap<String, AppInfo>,
     pub(crate) metadata: VecMap<String, String>,
@@ -86,7 +82,7 @@ pub(crate) struct AppInfo {
 pub(crate) struct VersionedName {
     /// A version name defaults at None, which means we need the latest version.
     pub(crate) version: Option<u64>,
-    /// The on-chain `Name` object that represents the dot_move name.
+    /// The on-chain `Name` object that represents the move registry name.
     pub(crate) name: Name,
 }
 
@@ -94,15 +90,16 @@ pub(crate) struct VersionedName {
 /// as we define it to deserialize on-chain data.
 #[derive(Debug, Serialize, Deserialize, Hash, Clone, Eq, PartialEq)]
 pub(crate) struct Name {
-    pub(crate) labels: Vec<String>,
-    pub(crate) normalized: String,
+    pub(crate) org: Domain,
+    pub(crate) app: Vec<String>,
 }
 
 impl Name {
-    pub(crate) fn new(app_name: &str, org_name: &str) -> Self {
-        let normalized = format!("{}@{}", app_name, org_name);
-        let labels = vec![org_name.to_string(), app_name.to_string()];
-        Self { labels, normalized }
+    pub(crate) fn new(org: Domain, app: &str) -> Self {
+        Self {
+            org,
+            app: vec![app.to_string()],
+        }
     }
 
     pub(crate) fn type_(package_address: SuiAddress) -> StructTag {
@@ -145,17 +142,20 @@ impl FromStr for VersionedName {
             return Err(MoveRegistryError::InvalidName(s.to_string()));
         };
 
-        let Some(app_name) = caps.get(1).map(|x| x.as_str()) else {
+        let Some(org_name) = caps.get(1).map(|x| x.as_str()) else {
             return Err(MoveRegistryError::InvalidName(s.to_string()));
         };
 
-        let Some(org_name) = caps.get(2).map(|x| x.as_str()) else {
+        // validate org_name by trying to cast our input to a Domain.
+        let domain = Domain::from_str(org_name)
+            .map_err(|_| MoveRegistryError::InvalidName(s.to_string()))?;
+
+        let Some(app_name) = caps.get(2).map(|x| x.as_str()) else {
             return Err(MoveRegistryError::InvalidName(s.to_string()));
         };
 
-        if (org_name.len() > MAX_LABEL_LENGTH) || (app_name.len() > MAX_LABEL_LENGTH) {
-            return Err(MoveRegistryError::InvalidName(s.to_string()));
-        };
+        // Validate our app's label.
+        validate_label(app_name).map_err(|_| MoveRegistryError::InvalidName(s.to_string()))?;
 
         let version: Option<u64> = caps
             .get(3)
@@ -165,7 +165,7 @@ impl FromStr for VersionedName {
 
         Ok(Self {
             version,
-            name: Name::new(app_name, org_name),
+            name: Name::new(domain, app_name),
         })
     }
 }
@@ -188,31 +188,36 @@ mod tests {
 
     #[test]
     fn parse_some_names() {
-        let versioned = VersionedName::from_str("app@org/v1").unwrap();
-        assert_eq!(versioned.name.normalized, "app@org");
+        let versioned = VersionedName::from_str("@org/app/1").unwrap();
         assert!(versioned.version.is_some_and(|x| x == 1));
 
-        assert!(VersionedName::from_str("app@org/v34")
+        assert!(VersionedName::from_str("@org/app/34")
             .unwrap()
             .version
             .is_some_and(|x| x == 34));
-        assert!(VersionedName::from_str("app@org")
+        assert!(VersionedName::from_str("@org/app")
             .unwrap()
             .version
             .is_none());
 
-        let ok_names = vec!["1-app@org/v1", "1-app@org/v34", "1-app@org"];
+        let ok_names = vec![
+            "@org/1-app/1",
+            "@org/1-app/34",
+            "@org/1-app",
+            "nested@org/app",
+            "even.more.nested@org/app",
+        ];
 
         let composite_ok_names = vec![
-            format!("{}@org/v1", generate_fixed_string(63)),
-            format!("{}-app@org/v34", generate_fixed_string(59)),
+            format!("@org/{}/1", generate_fixed_string(63)),
+            format!("@org/{}-app/34", generate_fixed_string(59)),
             format!(
-                "{}@{}",
+                "@{}/{}",
                 generate_fixed_string(63),
                 generate_fixed_string(63)
             ),
             format!(
-                "{}@{}-{}",
+                "@{}/{}-{}",
                 generate_fixed_string(63),
                 generate_fixed_string(30),
                 generate_fixed_string(30)
@@ -227,41 +232,41 @@ mod tests {
         }
 
         let not_ok_names = vec![
-            "-app@org",
-            "1.app@org",
-            "1--app@org",
-            "app-@org",
-            "app--@org",
-            "app@org/",
-            "app@org/v",
-            "app@org/veh",
+            "@org/-app",
+            "@org/1.app",
+            "@org/app-",
+            "@org/app--",
+            "@org/app?",
+            "@org/app/",
+            "@org/app/v",
+            "@org/app/veh",
             "@org",
-            "app@/veh",
+            "@/veh/app",
             "app",
             "@",
-            "ap@@org",
-            "ap!org",
-            "ap#org",
-            "ap#org@org",
-            "app%org",
+            "@org/ap@",
+            "!org/ap",
+            "#org/ap",
+            "@org/ap#org",
+            "org/@app",
             "",
             " ",
         ];
         let composite_err_names = vec![
             format!(
-                "{}--{}@{}",
+                "@{}{}--/{}",
                 generate_fixed_string(10),
                 generate_fixed_string(10),
                 generate_fixed_string(63)
             ),
             format!(
-                "--{}-{}@{}",
+                "@--{}-{}/{}",
                 generate_fixed_string(10),
                 generate_fixed_string(10),
                 generate_fixed_string(63)
             ),
             format!(
-                "{}@{}--{}",
+                "@{}/--{}{}",
                 generate_fixed_string(63),
                 generate_fixed_string(30),
                 generate_fixed_string(30)
