+++ crates/sui-graphql-rpc/tests/move_registry/move_registry/Move.toml
@@ -1,10 +1,10 @@
 [package]
-name = "dotmove"
+name = "move_registry"
 edition = "2024.beta"
 
 [dependencies]
 Sui = { local = "../../../../sui-framework/packages/sui-framework" }
 
 [addresses]
-dotmove = "0x0"
+move_registry = "0x0"
